+++ CHANGELOG.md
@@ -83,6 +83,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - alist: fix race condition while iterating [PR #1859]
 - build: enable building for Fedora40 and Ubuntu 24.04 [PR #1815]
 - ULC: build the universal linux client also for rpm based OS. [PR #1824]
+- consolidate: disable vfull duplicate job check [PR #1739]
 
 ### Removed
 - plugins: remove old deprecated postgres plugin [PR #1606]
@@ -181,6 +182,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [PR #1727]: https://github.com/bareos/bareos/pull/1727
 [PR #1728]: https://github.com/bareos/bareos/pull/1728
 [PR #1732]: https://github.com/bareos/bareos/pull/1732
+[PR #1739]: https://github.com/bareos/bareos/pull/1739
 [PR #1740]: https://github.com/bareos/bareos/pull/1740
 [PR #1744]: https://github.com/bareos/bareos/pull/1744
 [PR #1745]: https://github.com/bareos/bareos/pull/1745